fix(admin-service): TASK-BE-029-2-fix bootstrap token scope claim + SecureRandom bean
- BootstrapTokenService: add scope claim to issue(); verifyAndConsume(token, requiredScope) enforces scope membership; default scope set {2fa_enroll, 2fa_verify}.
- BootstrapAuthenticationFilter: map path -> required scope (enroll -> 2fa_enroll, verify -> 2fa_verify); mismatch yields 401 INVALID_BOOTSTRAP_TOKEN.
- AdminLoginService: explicitly mint bootstrap with both scopes on enrollment-required.
- TotpEnrollmentService: drop static SecureRandom; inject via constructor from new CryptoConfig bean; separate OperatorNotFoundException from AuditFailureException.
- AdminOperatorTotpJpaEntity.version: long -> int to match operator convention (DDL remains BIGINT).
- Tests: scope-positive/negative in BootstrapTokenServiceTest; deterministic SHA1PRNG recovery-code test in new TotpEnrollmentServiceTest; AdminAuthControllerTest stub matches new 2-arg verifyAndConsume.

+++ apps/admin-service/src/main/java/com/example/admin/infrastructure/persistence/AdminOperatorTotpJpaEntity.java
@@ -43,7 +43,7 @@ public class AdminOperatorTotpJpaEntity {
 
     @Version
     @Column(name = "version", nullable = false)
-    private long version;
+    private int version;
 
     public static AdminOperatorTotpJpaEntity create(long operatorId,
                                                     byte[] secretEncrypted,
